Set proposalJ to have RC and rcJ to have prepares. Propagate changes
