+++ qbft/spectest/tests/messages/create_proposal_prev_prepared.go
@@ -14,16 +14,16 @@ func CreateProposalPreviouslyPrepared() tests.SpecTest {
 		CreateType: tests.CreateProposal,
 		Name:       "create proposal previously prepared",
 		Value:      [32]byte{1, 2, 3, 4},
-		RoundChangeJustifications: []*qbft.SignedMessage{
+		ProposalJustifications: []*qbft.SignedMessage{
 			testingutils.TestingRoundChangeMessageWithRound(ks.Shares[1], types.OperatorID(1), 2),
 			testingutils.TestingRoundChangeMessageWithRound(ks.Shares[2], types.OperatorID(2), 2),
 			testingutils.TestingRoundChangeMessageWithRound(ks.Shares[3], types.OperatorID(3), 2),
 		},
-		ProposalJustifications: []*qbft.SignedMessage{
+		RoundChangeJustifications: []*qbft.SignedMessage{
 			testingutils.TestingPrepareMessage(ks.Shares[1], types.OperatorID(1)),
 			testingutils.TestingPrepareMessage(ks.Shares[2], types.OperatorID(2)),
 			testingutils.TestingPrepareMessage(ks.Shares[3], types.OperatorID(3)),
 		},
-		ExpectedRoot: "282dd7899470e882fc22e9284628f4c25b2e3ba89bc0f50becb677c9a2e4708c",
+		ExpectedRoot: "8d0611de4069f8fab12af6fc13907567c39b999071773e0891631bdc15dc17ac",
 	}
 }
