+++ qbft/spectest/tests/messages/create_round_change_prev_prepared.go
@@ -11,14 +11,14 @@ import (
 func CreateRoundChangePreviouslyPrepared() tests.SpecTest {
 	ks := testingutils.Testing4SharesSet()
 	return &tests.CreateMsgSpecTest{
-		CreateType: tests.CreateProposal,
+		CreateType: tests.CreateRoundChange,
 		Name:       "create round change previously prepared",
 		Value:      [32]byte{1, 2, 3, 4},
-		ProposalJustifications: []*qbft.SignedMessage{
+		RoundChangeJustifications: []*qbft.SignedMessage{
 			testingutils.TestingPrepareMessage(ks.Shares[1], types.OperatorID(1)),
 			testingutils.TestingPrepareMessage(ks.Shares[2], types.OperatorID(2)),
 			testingutils.TestingPrepareMessage(ks.Shares[3], types.OperatorID(3)),
 		},
-		ExpectedRoot: "a095b2f42e32e62980d9675cfcf3a58ed4da4d05b63463e182a451461de3b44e",
+		ExpectedRoot: "e08923100c9970e713651b4ff2ce0308b0c761e97eb009b93beefc0b923828f5",
 	}
 }
