+++ qbft/spectest/tests/messages/invalid_proposal_justification_unmarshaling.go
@@ -2,7 +2,7 @@ package messages
 
 import "github.com/bloxapp/ssv-spec/qbft/spectest/tests"
 
-// InvalidProposalJustificationsUnmarshalling tests unmarshalling invalid prepare justifications (during message.validate())
+// InvalidProposalJustificationsUnmarshalling tests unmarshalling invalid rc messages justification (during message.validate())
 func InvalidProposalJustificationsUnmarshalling() tests.SpecTest {
 	panic("implement")
 }
