+++ qbft/spectest/tests/messages/invalid_round_change_justification_unmarshaling.go
@@ -2,7 +2,7 @@ package messages
 
 import "github.com/bloxapp/ssv-spec/qbft/spectest/tests"
 
-// InvalidRoundChangeJustificationsUnmarshalling tests unmarshalling invalid round change justifications (during message.validate())
+// InvalidRoundChangeJustificationsUnmarshalling tests unmarshalling invalid prepare messages justification (during message.validate())
 func InvalidRoundChangeJustificationsUnmarshalling() tests.SpecTest {
 	panic("implement")
 }
