+++ qbft/spectest/tests/messages/proposal_justification_unmarshaling.go
@@ -2,7 +2,7 @@ package messages
 
 import "github.com/bloxapp/ssv-spec/qbft/spectest/tests"
 
-// ProposalJustificationsUnmarshalling tests unmarshalling prepare justifications
+// ProposalJustificationsUnmarshalling tests unmarshalling rc messages justification
 func ProposalJustificationsUnmarshalling() tests.SpecTest {
 	panic("implement")
 }
