+++ qbft/spectest/tests/messages/rc_justification_unmarshaling.go
@@ -2,7 +2,7 @@ package messages
 
 import "github.com/bloxapp/ssv-spec/qbft/spectest/tests"
 
-// RoundChangeJustificationsUnmarshalling tests unmarshalling round change justifications
+// RoundChangeJustificationsUnmarshalling tests unmarshalling prepare justifications
 func RoundChangeJustificationsUnmarshalling() tests.SpecTest {
 	panic("implement")
 }
