+++ qbft/spectest/tests/proposal/prepared_previously_no_prop_justification_quorum.go
@@ -1,0 +1,39 @@
+package proposal
+
+import (
+	"github.com/bloxapp/ssv-spec/qbft"
+	"github.com/bloxapp/ssv-spec/qbft/spectest/tests"
+	"github.com/bloxapp/ssv-spec/types"
+	"github.com/bloxapp/ssv-spec/types/testingutils"
+)
+
+// PreparedPreviouslyNoProposalJustificationQuorum tests a proposal for > 1 round, prepared previously but without quorum of round change msgs justification
+func PreparedPreviouslyNoProposalJustificationQuorum() tests.SpecTest {
+	pre := testingutils.BaseInstance()
+
+	ks := testingutils.Testing4SharesSet()
+	prepareMsgs := []*qbft.SignedMessage{
+		testingutils.TestingPrepareMessage(ks.Shares[1], types.OperatorID(1)),
+		testingutils.TestingPrepareMessage(ks.Shares[2], types.OperatorID(2)),
+		testingutils.TestingPrepareMessage(ks.Shares[3], types.OperatorID(3)),
+	}
+	rcMsgs := []*qbft.SignedMessage{
+		testingutils.TestingRoundChangeMessageWithRound(ks.Shares[1], types.OperatorID(1), 2),
+		testingutils.TestingRoundChangeMessageWithRound(ks.Shares[2], types.OperatorID(2), 2),
+	}
+
+	msgs := []*qbft.SignedMessage{
+		testingutils.TestingProposalMessageWithParams(ks.Shares[1], types.OperatorID(1), 2, qbft.FirstHeight,
+			testingutils.TestingQBFTRootData,
+			testingutils.MarshalJustifications(rcMsgs), testingutils.MarshalJustifications(prepareMsgs),
+		),
+	}
+	return &tests.MsgProcessingSpecTest{
+		Name:           "no rc quorum (prepared)",
+		Pre:            pre,
+		PostRoot:       "5b18ca0b470208d8d247543306850618f02bddcbaa7c37eb6d5b36eb3accb5fb",
+		InputMessages:  msgs,
+		OutputMessages: []*qbft.SignedMessage{},
+		ExpectedError:  "invalid signed message: proposal not justified: change round has no quorum",
+	}
+}
