--- qbft/spectest/tests/proposal/prepared_previously_no_proposal_justification_quorum.go
@@ -1,51 +1,0 @@
-package proposal
-
-import (
-	"github.com/bloxapp/ssv-spec/qbft"
-	"github.com/bloxapp/ssv-spec/qbft/spectest/tests"
-	"github.com/bloxapp/ssv-spec/types"
-	"github.com/bloxapp/ssv-spec/types/testingutils"
-	"github.com/bloxapp/ssv-spec/types/testingutils/comparable"
-)
-
-// PreparedPreviouslyNoProposalJustificationQuorum tests a proposal for > 1 round, prepared previously but without quorum of prepared msgs justification
-func PreparedPreviouslyNoProposalJustificationQuorum() tests.SpecTest {
-	pre := testingutils.BaseInstance()
-	ks := testingutils.Testing4SharesSet()
-	sc := preparedPreviouslyNoProposalJustificationQuorumStateComparison()
-
-	prepareMsgs := []*qbft.SignedMessage{
-		testingutils.TestingPrepareMessage(ks.Shares[1], types.OperatorID(1)),
-		testingutils.TestingPrepareMessage(ks.Shares[2], types.OperatorID(2)),
-	}
-	rcMsgs := []*qbft.SignedMessage{
-		testingutils.TestingRoundChangeMessageWithRoundAndRC(ks.Shares[1], types.OperatorID(1), 2,
-			testingutils.MarshalJustifications(prepareMsgs)),
-		testingutils.TestingRoundChangeMessageWithRoundAndRC(ks.Shares[2], types.OperatorID(2), 2,
-			testingutils.MarshalJustifications(prepareMsgs)),
-		testingutils.TestingRoundChangeMessageWithRoundAndRC(ks.Shares[3], types.OperatorID(3), 2,
-			testingutils.MarshalJustifications(prepareMsgs)),
-	}
-
-	msgs := []*qbft.SignedMessage{
-		testingutils.TestingProposalMessageWithParams(ks.Shares[1], types.OperatorID(1), 2, qbft.FirstHeight,
-			testingutils.TestingQBFTRootData,
-			testingutils.MarshalJustifications(rcMsgs), testingutils.MarshalJustifications(prepareMsgs),
-		),
-	}
-	return &tests.MsgProcessingSpecTest{
-		Name:           "no prepare quorum (prepared)",
-		Pre:            pre,
-		PostRoot:       sc.Root(),
-		PostState:      sc.ExpectedState,
-		InputMessages:  msgs,
-		OutputMessages: []*qbft.SignedMessage{},
-		ExpectedError:  "invalid signed message: proposal not justified: change round msg not valid: no justifications quorum",
-	}
-}
-
-func preparedPreviouslyNoProposalJustificationQuorumStateComparison() *comparable.StateComparison {
-	state := testingutils.BaseInstance().State
-
-	return &comparable.StateComparison{ExpectedState: state}
-}
