+++ qbft/spectest/tests/proposal/prepared_previously_no_rc_justification_quorum.go
@@ -5,21 +5,26 @@ import (
 	"github.com/bloxapp/ssv-spec/qbft/spectest/tests"
 	"github.com/bloxapp/ssv-spec/types"
 	"github.com/bloxapp/ssv-spec/types/testingutils"
+	"github.com/bloxapp/ssv-spec/types/testingutils/comparable"
 )
 
-// PreparedPreviouslyNoRCJustificationQuorum tests a proposal for > 1 round, prepared previously but without quorum of round change msgs justification
-func PreparedPreviouslyNoRCJustificationQuorum() tests.SpecTest {
+// PreparedPreviouslyNoRoundChangeJustificationQuorum tests a proposal for > 1 round, prepared previously but without quorum of prepared msgs justification
+func PreparedPreviouslyNoRoundChangeJustificationQuorum() tests.SpecTest {
 	pre := testingutils.BaseInstance()
-
 	ks := testingutils.Testing4SharesSet()
+	sc := preparedPreviouslyNoRoundChangeJustificationQuorumStateComparison()
+
 	prepareMsgs := []*qbft.SignedMessage{
 		testingutils.TestingPrepareMessage(ks.Shares[1], types.OperatorID(1)),
 		testingutils.TestingPrepareMessage(ks.Shares[2], types.OperatorID(2)),
-		testingutils.TestingPrepareMessage(ks.Shares[3], types.OperatorID(3)),
 	}
 	rcMsgs := []*qbft.SignedMessage{
-		testingutils.TestingRoundChangeMessageWithRound(ks.Shares[1], types.OperatorID(1), 2),
-		testingutils.TestingRoundChangeMessageWithRound(ks.Shares[2], types.OperatorID(2), 2),
+		testingutils.TestingRoundChangeMessageWithRoundAndRC(ks.Shares[1], types.OperatorID(1), 2,
+			testingutils.MarshalJustifications(prepareMsgs)),
+		testingutils.TestingRoundChangeMessageWithRoundAndRC(ks.Shares[2], types.OperatorID(2), 2,
+			testingutils.MarshalJustifications(prepareMsgs)),
+		testingutils.TestingRoundChangeMessageWithRoundAndRC(ks.Shares[3], types.OperatorID(3), 2,
+			testingutils.MarshalJustifications(prepareMsgs)),
 	}
 
 	msgs := []*qbft.SignedMessage{
@@ -29,11 +34,18 @@ func PreparedPreviouslyNoRCJustificationQuorum() tests.SpecTest {
 		),
 	}
 	return &tests.MsgProcessingSpecTest{
-		Name:           "no rc quorum (prepared)",
+		Name:           "no prepare quorum (prepared)",
 		Pre:            pre,
-		PostRoot:       "5b18ca0b470208d8d247543306850618f02bddcbaa7c37eb6d5b36eb3accb5fb",
+		PostRoot:       sc.Root(),
+		PostState:      sc.ExpectedState,
 		InputMessages:  msgs,
 		OutputMessages: []*qbft.SignedMessage{},
-		ExpectedError:  "invalid signed message: proposal not justified: change round has no quorum",
+		ExpectedError:  "invalid signed message: proposal not justified: change round msg not valid: no justifications quorum",
 	}
+}
+
+func preparedPreviouslyNoRoundChangeJustificationQuorumStateComparison() *comparable.StateComparison {
+	state := testingutils.BaseInstance().State
+
+	return &comparable.StateComparison{ExpectedState: state}
 }
