+++ qbft/spectest/tests/roundchange/invalid_justification_full_data.go
@@ -2,7 +2,7 @@ package roundchange
 
 import "github.com/bloxapp/ssv-spec/qbft/spectest/tests"
 
-// InvalidJustificationFullData tests round change justification for which H(full data) != root
+// InvalidJustificationFullData tests proposal justification for which H(full data) != root
 func InvalidJustificationFullData() tests.SpecTest {
 	panic("implement")
 }
